requirements.txt
asgiref==3.8.1
Django==5.1.6
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
mysqlclient==2.2.7
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
